employee_mgmt_app/server/requirements.txt
aniso8601==9.0.1
attrs==23.1.0
blinker==1.6.2
cffi==1.15.1
click==8.1.7
console-logging==0.0.5.0
cryptography==41.0.3
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JSON==0.4.0
Flask-MySQL==1.5.2
Flask-RESTful==0.3.10
Flask-Migrate==2.6.0
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
pytz==2023.3.post1
referencing==0.30.2
rpds-py==0.10.2
six==1.16.0
SQLAlchemy==2.0.20
termcolor==2.3.0
typing_extensions==4.7.1
Werkzeug==2.3.7
python-dotenv
gunicorn
